requirements.txt
customtkinter>=5.2.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
reportlab>=4.0.0
tkinterdnd2>=0.3.0
